rename: MyChannelInfoUpdateService
