+++ src/main/java/com/outsider/masterofpredictionbackend/user/command/domain/embeded/Gender.java
@@ -1,0 +1,6 @@
+package com.outsider.masterofpredictionbackend.user.command.domain.embeded;
+
+public enum Gender {
+    MALE,
+    FEMALE,
+}
