+++ src/main/java/com/outsider/masterofpredictionbackend/user/command/domain/embeded/Location.java
@@ -1,0 +1,8 @@
+package com.outsider.masterofpredictionbackend.user.command.domain.embeded;
+
+public enum Location {
+    KOREA,
+    USA,
+    JAPAN,
+    CHINA
+}
