--- src/main/java/com/outsider/masterofpredictionbackend/user/command/domain/model/Gender.java
@@ -1,7 +1,0 @@
-package com.outsider.masterofpredictionbackend.user.command.domain.model;
-
-public enum Gender {
-    MALE,
-    FEMALE,
-    TRANSGENDER
-}
