--- src/main/java/com/outsider/masterofpredictionbackend/user/command/domain/model/Location.java
@@ -1,8 +1,0 @@
-package com.outsider.masterofpredictionbackend.user.command.domain.model;
-
-public enum Location {
-    KOREA,
-    USA,
-    JAPAN,
-    CHINA
-}
